fix: isolate stable layout memo helper
