+++ packages/app/src/pages/session/session-layout.ts
@@ -1,6 +1,7 @@
 import { useParams } from "@solidjs/router"
 import { createMemo } from "solid-js"
 import { useLayout } from "@/context/layout"
+import { createStableLayoutMemo } from "./stable-layout-memo"
 
 export const useSessionKey = () => {
   const params = useParams()
@@ -16,22 +17,5 @@ export const useSessionLayout = () => {
     sessionKey,
     tabs: createStableLayoutMemo(() => layout.tabs(sessionKey)),
     view: createStableLayoutMemo(() => layout.view(sessionKey)),
-  }
-}
-
-export function createStableLayoutMemo<T>(read: () => T) {
-  let last: { value: T } | undefined
-  const memo = createMemo(read)
-
-  return () => {
-    const value = memo()
-    if (value !== undefined) {
-      last = { value }
-      return value
-    }
-
-    if (last) return last.value
-
-    throw new Error("Stable layout memo read before initialization")
   }
 }
